requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Environment
python-dotenv==1.0.0
gunicorn==22.0.0
gevent==24.2.1
# HTTP Client
requests==2.31.0

# Core utilities
numpy==1.24.3

# ML - CPU optimized versions (smaller download)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
